Created methods for Nodes Creation & Added Desc. Comments
Added Descriptive comments for the Constructors used inside the CustomNode.java.

The code for creating Number and Operator nodes are transferred to their own methods.

I think the converter function is harder to read when the code for creating the CustomNode is inside.

The purpose and intent of the code is different from "converting" and is more oriented for "creating".