Created a recursive loop to allow user to enter input again
I created a function that will handle the receiving input fromt the user and will be called after the execution of the programStart method.

Depending on user input, the function will call the programStart method or end the function without calling for any other method.

Added a method to warn user of wrong input.

When user does not follow the post fix notation or does not input an operator with their input.